Principal changes:
An alternative to PARTEL_PARA written in python. As the Fortran version seemed to be "hanging" a lot with a lot of memory required, this was done specifically for 2D SELAFIN files (to solve our problem on a deadline critical project).

Important note 1: this python version does not need to know about VOISIN. The ideas it contains may be useful to its Fortran counterpart.
Important note 2: this version need to follow a call to PARTEL_PRELIM. Because of issue with the install of METIS on Windows, this python version may eventually replace METIS as well.

This new development can now also be activated through the python scripts. By default, if the key par_cmdexec is not in the systel.cfg, or if it is:
par_cmdexec:   <config>/partel >> <partel.log>
... then the default PARTEL is used.

If the same key set to:
par_cmdexec:   <config>\partel_prelim.exe; <mpi_cmdexec> <config>\partel_para.exe >> <partel.log>
... then the new Fortran PARTEL in parallel is used.

If the same key set to:
par_cmdexec:   <config>\partel_prelim.exe; python <root>\pytel\utils\partitioning.py >> <partel.log>
... then the new Python partitioning (in scalar so far) is used.

/!\ WARNING: so far this is only working on Linux because of an MPI error.

==============================
Side changes: the ".cmd" batch files are being removed to make a distinction between the script in the <root>\bin and those in <root>\pytel.

git-svn-id: http://svn.opentelemac.org/svn/opentelemac/trunk@1143 21fd64ea-cbdc-4a4a-8fb2-7ac88764eefb

Former-commit-id: ee13ae3cf8ff8dd4a565ffd4f619be074e8ad600