Script Python -- dealing with SELAFIN files:
Additional capability implemented for all SLF objects when used to process multiple simulation results (when produced by TELEMAC in parallel mode). The input will be the
 root of all files, and the output is re-collected into one.

By extension runSELAFIN.py merge --parallel duplicates what is done in GRETEL -- it was not intentional.

Option --parallel is required to activate this feature. Note that doing operation on one sub-file only, to produce one sub-file is not covered here as it could be done already.

git-svn-id: http://svn.opentelemac.org/svn/opentelemac/trunk@1234 21fd64ea-cbdc-4a4a-8fb2-7ac88764eefb

Former-commit-id: 0bc43f610d789de304bc9a05732f7de829d78ce7