[python][cue feature #223, validated through guppy] Compilation (compileTELEMAC.py) is now done in parallel, where the parallelisation is for all independent *.o and all packaging *.lib, still, one module at a time, and one configuration at a time. Compilation now takes up all processors available on a computer and only those just to make sure that the system does not overload.
That is all I can possible think of doing at the moment. Anything else would have to be done in the background without screen printout progress or more likely, running independent compileTELEMAC -m ... -c ... for all modules and configuration. But since all processors are now busy anyway, there is not much interest in doing so.



git-svn-id: http://svn.opentelemac.org/svn/opentelemac/trunk@5866 21fd64ea-cbdc-4a4a-8fb2-7ac88764eefb

Former-commit-id: 8afc928d9aab47f9fb2d6e2a918ca85d0ab843de